Fixing the BVT Issues:
1) self-heal failures: With the recent changes made to gluster for the bug: https://bugzilla.redhat.com/show_bug.cgi?id=1480423, the location of the brick process pid's changed to /var/run/gluster.
Making the corresponding changes to glusto-tests libraries.
Moving away from referring to pid file to grep for the process with the brick name.
This fixes the issue.

2) Group options not being set properly: Since we were popping the 'group' option from the 'options'
dictionary after the group options being set to set the other volume options, the option gets removed
from the g.config['gluster']['smb_volume_options'] as well.
Hence perform a deep copy of the dict before modifying the dict.

Change-Id: I293bf81913857cb0327f30aa1db5aaa9be5a318e
Signed-off-by: ShwethaHP <spandura@redhat.com>