functional/disperse: verify IO hang during client side heal
When the IOs are done with server side heal disabled,
it should not hang.

ec_check_heal_comp function will fail because of the
bug 1593224-  Client side heal is not removing dirty
flag for some of the files
While this bug has been raised and investigated by
dev, this patch is doing its job and testing the
target functionality.

RHG3-11097
Signed-off-by: Sunil Kumar Acharya <sheggodu@redhat.com>

Signed-off-by: Ashish Pandey <aspandey@redhat.com>
Change-Id: I841285c9b1a747f5800ec8cdd29a099e5fcc08c5
Signed-off-by: Ashish Pandey <aspandey@redhat.com>