Changing error messages to be checked as per new messages.
Changing error message displayed when peer detach is issued
with bricks are present on the node which is being detached.
Adding a logic to handle both the new as well as the old
error message.

Old msg:
peer detach: failed: Brick(s) with the peer <my_server>
exist in cluster

New msg:
peer detach: failed: Peer <my_server> hosts one or more bricks.
If the peer is in not recoverable state then use either
replace-brick or remove-brick command with force to remove
all bricks from the peer and attempt the peer detach again.

Change-Id: I3d8fdac2c33638ecc2a8b5782c68caebbf17cf41
Signed-off-by: kshithijiyer <kshithij.ki@gmail.com>