Adding missing LICENSE file to the project.
For all opensouce projects it's a must to have a
LICENSE file in the root directory of the project.
The LICENSE file describes the terms and conditions
under which the project is open sourced. glusto-tests
follows GNU GPL LICENSE which is missing this file
in it's root directory. Hence adding GNU GPL
LICENSE file to the project.

Change-Id: I7e3675d1b7540c42cc0549aa501f15585887ed12
Signed-off-by: kshithijiyer <kshithij.ki@gmail.com>