Add the contributor guide
This guide provides extensive steps to get the developer
started with the development for the project.
There needs to be a contributor guide for every project.

Things like git and gerrit setup which are essential for
the development systems have been included in this guide
along with some good practices where applicable.

Change-Id: Ie1ad6e08680f2c30f9361a20c2ba68493abfb054
Signed-off-by: Vinayak Papnoi <vpapnoi@redhat.com>