Add library to reset particular volume option
Currently there is a library to reset the volume entirely.
There needs to be a library to reset a particular volume option
because if a user has set a lot of other volume options then it
becomes a tedious job to reset and set the other options explicitly.

The library function 'reset_volume_option' takes particular
volume options as a string and resets it.

Change-Id: I2049ea4a49d321d921d940f34e124d1f34bfa068
Signed-off-by: Vinayak Papnoi <vpapnoi@redhat.com>