Fix TypeError caused in heal_libs.py
The line 316 of heal_libs.py library contains a function called
'wait_for_self_heal_daemons_to_be_online' which makes use of the
'volname' and 'timeout' arguements for string formatting.

The method used for the string formatting was wrong which causes
an error 'TypeError: not enough arguments for format string'.
Due to this, all the test cases which make use of heal_libs.py
will result in a TypeError scenario.

This patch will fix the TypeError for the string formatting.

Change-Id: I73abe573d2ccb60b74ccd6ae268b950640d75600
Signed-off-by: Vinayak Papnoi <vpapnoi@redhat.com>