[testfix] Update test_snap_list_after_restart
There can be a case where due to a failed tearDown of a previous
test case, the snapshot creation might fail. This can happen when
a redundant snapshot with the same name is present.

This fix includes the changed snapshot names along with tearDown
for deleting the snapshots created in the test case
"test_snap_list_after_restart.py".

This fix also includes the method 'wait_for_glusterd_to_start'
where there is a restart of glusterd.
Adding a check for validation of snapshots using snapname.

Change-Id: If8b48a12bd067ad54dba742eeb88444beaf5f153
Signed-off-by: Vinayak Papnoi <vpapnoi@redhat.com>