[lib] Adding geo-rep params in gluster_base_class.py
Adding params in gluster_base_class.py to store value
of root user's password and non-root user's name, group
and password. sync_method will be mentioned inside the
testcase as there are testcases which are only specific
to one of the sync type and it doesn't make sense to add
it in the config file. Instead we'll write testcase as shown
below:
```
class georeptestcase(baseclass):

     run_chmod_with_changelog_crawl(sync_type):
        # Steps of the testcase

     test_chmod_with_changelog_crawl_rsync():
        # Calling function with rsync as sync_type.
        run_chmod_with_changelog_crawl('rsync')

     test_chmod_with_changelog_crawl_tarssh():
        # Calling function with tarssh as sync_type.
        run_chmod_with_changelog_crawl('tarssh')
```

Change-Id: Ie65f542e76bfbee89ac2914bdcd086e1bd08dfdb
Signed-off-by: kshithijiyer <kshithij.ki@gmail.com>