[Libfix] Change daemon_reload() to support newer platforms
Problem:
Currently the code supports both service and systemctl
commands but it fails on the latest platforms with
the below error on the latest platforms:
```
service glusterd reload
Redirecting to /bin/systemctl reload glusterd.service
Failed to reload glusterd.service: Job type reload is
not applicable for unit glusterd.service.
```

This is because the latest platforms uses systemctl
instead of service to reload the daemon processes:
```
systemctl daemon-reload
```

Solution:
The present code doesn't work properly as the check
is specific to only one platform, hence it fails.
The solution for this is to just check for older
platforms and run service command. For all other
platforms run systemctl command.

Change-Id: I19b24652b96c4794553d3659eaf0301395929bca
Signed-off-by: kshithijiyer <kshithij.ki@gmail.com>