[TestFix] Fix to make TC compatible with DHT lib
The TC was failing with "AssertionError: ('hash range is not
there %s', False)" even though the bricks were healed and the
directory was created on non-hashed bricks. This was due to the
conflict between the TC and the DHT library changes (added to
fix the issues caused by DHT pass-through functionality). The
code is now modified according to the library changes and hence
the TC works fine.

Change-Id: I501e7db89643822fbc711e631ceacda79e4c4ea4
Signed-off-by: sayaleeraut <saraut@redhat.com>