[Test] Add tc to check get-state when brick is killed
Test case verifies whether the gluster get-state shows the proper brick status in the output.
The test case checks the brick status when the brick is up and also after killing the brick process.
It also verifies whether the other bricks are up when a particular brick process is killed.

Change-Id: I9801249d25be2817104194bb0a8f6a16271d662a
Signed-off-by: Pranav <prprakas@redhat.com>