[Testfix] Fix TypeError for python3
Problem: Test is failing with below traceback
when ran with python3 as default.
`
Traceback (most recent call last):
  File "<string>", line 1, in <module>
TypeError: a bytes-like object is required, not 'str'
`
Solution:
Added ''.encode() which will fix the issue when ran
using both python2 and python3

Added a check for core file on the client node.

Change-Id: I8f800f5fad97c3b7591db79ea51203e5293a1f69
Signed-off-by: Bala Konda Reddy M <bala12352@gmail.com>