[Testfix] Increase sleep time in test_alert_time_out
Problem:
Test script test_alert_time_out currently fails
2 out of 6 times when executed on the same setup
this is due to the log files not have 120004 A
alert message. This issue is only observed in
distributed volume type mounted over fuse
protocol.

Solution:
There is no permanent solution to this problem
as even if we increase the sleep 20 seconds there
is still a chance that it might fail. The optimal
sleep time where it only fails 5 times after 15
attempts is 6 seconds. Hence changing sleep time
to 6 seconds.

Change-Id: I9e9bd41321e24f502d90c3c34edce9113133755e
Signed-off-by: kshithijiyer <kshithij.ki@gmail.com>