[Config] Add ISSUE_TEMPLATE for github issues
Problem:
There have been a couple of instances where we have received
issues which don't have a clear description,
making it difficult for glusto-tests developers to fix it.
One such example is [1].

Solution:
Adding a ISSUE_TEMPLATE for github issues
so that the reporter of the issue follows
a given format for the issue.

Links:
[1] https://github.com/gluster/glusto-tests/issues/19

Fixes: https://github.com/gluster/glusto-tests/issues/29
Change-Id: I0d3265ccc373a919d5d4fc7cc8df283f8306c4a5
Signed-off-by: kshithijiyer <kshithij.ki@gmail.com>