[LibFix] Add encoding type to get_fattr()
Currently the get_fattr() function returns the xattr value in the
hex encoding format. Adding the option to add/specify other
encoding types, i.e. text and base64, where hex will be the default
encoding format.

Reason -
When the xattrs are custom set through mountpoint, it becomes
easier to test if the value is correct by using the "text"
encoding.

Example -
>>> ret = get_fattr(host, fqpath, fattr)
>>> print ret
0x414243

>>> ret = get_fattr(host, fqpath, fattr, encode="text")
>>> print ret
"ABC"

The value "ABC" is easily readable as opposed to 0x414243 when
performing a test.

Change-Id: Ie2377b924816ebab0a2af116d82600e01f03d61f
Signed-off-by: sayaleeraut <saraut@redhat.com>