[Testfix] Remove redundant logging - Part 2
Problem:
In most of the testcases due to redundant logging,
the performance of the whole suite completion time
is affected.

Solution:
In BVT Test suite, there are 184 g.log.info messages
more than half of them are redundant.
Removed logs wherever it is not required.

Added missing get_super_method for setUp and tearDown
for one testcase, modified increment in the test.

Change-Id: I19e4462f2565906710c2be117bc7c16c121ddd32
Signed-off-by: Bala Konda Reddy M <bala12352@gmail.com>