[Libfix] Add interval_check for monitor_heal_completion
Adding a parameter `interval_check` will ease the user and help
in the reducing the waiting time for heal in scenarios.
By default, ls -l <brickpath> | grep -ve "xattrop-" | wc -l
is checked for every 2 minutes.

Problem: I have 100 files that needs to be healed, after 2 minutes.
suppose there are only 2/3 files that needs to be healed.
With the existing approach the next check will wait for whole 2
minutes even though the files would have healed by 10 seconds after
previouscheck.

Solution: Giving an option for the user to check at which interval
to check for the files that needs to healed, we can reduce the
unnecssary waiting time.

It won't affect the existing cases as the interval_check is defaults
to 120.

Change-Id: Ib288c75549644b6f6c94b5288f1c07cce7933915
Signed-off-by: Bala Konda Reddy M <bala12352@gmail.com>