[TestFix] Fix test_sosreport_interoperability.py
Fixing the comaprison logic for dirs before SOS report
generation and after

Problem:
The current approach is using `==` operator for comparing
two lists. The issue happens when the lists are in different
order,
for e.g:
>>>a = [1,2,3]
>>>b = [1,3,2]
>>>a == b
False

The comparison returns False, whereas the content of the lists
are same.

Solution:

Use set() instead to ensure that the contents are similar,
E.g.
>>> set(a) - set(b)
set()

Signed-off-by: Pranav <prprakas@redhat.com>