[TestFix] Fix test_metadata_self_heal_on_open_fd
Problem:
Method "get_heal_info_summary" returns dict with 'hostnames' as
keys. and in the test  we are using 'bricks_list' items, which
contains ip addr for accessing the above dict. This causes the
test to fail with 'KeyError' in CI env

Solution:
Change the logic to append hostnames in 'bricks_list' instead of
hostnames

Signed-off-by: Pranav <prprakas@redhat.com>